avr_phm/requirements.txt
torch==2.5.0
torchvision==0.20.0
deepxde==1.11.2
scipy==1.14.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
shap==0.46.0
wandb==0.18.5
matplotlib==3.9.2
seaborn==0.13.2
pyyaml==6.0.2
tqdm==4.66.5
imbalanced-learn==0.12.3
pytest==8.3.3
pytorch-lightning==2.4.0
foolbox==3.3.3
torchinfo==1.8.0
fvcore==0.1.5.post20221221
